backend/requirements.txt
Flask==2.3.2
Flask-CORS==3.0.10
mysql-connector-python==8.0.33
python-dotenv==1.0.0